Reorganize vector data importer. Rename custom_tools to vector_data_importer and move UI code from custom_widgets to vector_data_importer
