+++ tests/test_vector_data_importer/data/connection_nodes.gpkg
