+++ tests/test_vector_data_importer/data/import_connection_nodes.json
