+++ tests/test_vector_data_importer/data/import_weirs_snap.json
