+++ tests/test_vector_data_importer/data/integrate_weirs_snap.json
