+++ tests/test_vector_data_importer/data/originals/schematisation_channel_with_weir.gpkg
