+++ tests/test_vector_data_importer/data/weirs.gpkg
