+++ tests/test_vector_data_importer/data/weirs_too_long.gpkg
