+++ threedi_schematisation_editor/vector_data_importer/dialogs/attribute_value_map.py
@@ -4,7 +4,7 @@ from qgis.core import NULL
 from qgis.PyQt.QtWidgets import QInputDialog, QTableWidgetItem
 
 from threedi_schematisation_editor.utils import QUOTED_NULL
-from threedi_schematisation_editor.custom_widgets import vm_basecls, vm_uicls
+from threedi_schematisation_editor.vector_data_importer.dialogs import vm_basecls, vm_uicls
 
 
 class AttributeValueMapDialog(vm_basecls, vm_uicls):
