+++ threedi_schematisation_editor/vector_data_importer/utils.py
